Refactor setting build metadata
