+++ bundler/spec/support/helpers.rb
@@ -330,13 +330,7 @@ module Spec
 
         replace_version_file(version, dir: build_path) # rubocop:disable Style/HashSyntax
 
-        build_metadata = {
-          :built_at => loaded_gemspec.date.utc.strftime("%Y-%m-%d"),
-          :git_commit_sha => git_commit_sha,
-          :release => true,
-        }
-
-        replace_build_metadata(build_metadata, dir: build_path) # rubocop:disable Style/HashSyntax
+        Spec::BuildMetadata.write_build_metadata(dir: build_path) # rubocop:disable Style/HashSyntax
 
         gem_command "build #{relative_gemspec}", :dir => build_path
