+++ bundler/task/release.rake
@@ -1,26 +1,17 @@
 # frozen_string_literal: true
 
 require_relative "../lib/bundler/gem_tasks"
+require_relative "../spec/support/build_metadata"
 
 Bundler::GemHelper.tag_prefix = "bundler-"
 
 task :build_metadata do
-  build_metadata = {
-    :built_at => Bundler::GemHelper.gemspec.date.utc.strftime("%Y-%m-%d"),
-    :git_commit_sha => `git rev-parse --short HEAD`.strip,
-    :release => true,
-  }
-
-  Spec::Path.replace_build_metadata(build_metadata)
+  Spec::BuildMetadata.write_build_metadata
 end
 
 namespace :build_metadata do
   task :clean do
-    build_metadata = {
-      :release => false,
-    }
-
-    Spec::Path.replace_build_metadata(build_metadata)
+    Spec::BuildMetadata.reset_build_metadata
   end
 end
